Particles - Instancing

+++ DirectHII/d_particles.h
@@ -1,6 +1,6 @@
 
 // particles
 void D_BeginParticles (void);
-void D_AddParticle (float *origin, unsigned color, int size);
+void D_AddParticle (float *origin, unsigned color, float size);
 void D_EndParticles (void);
